refactor/refatorar o método delete para retornar um booleano

+++ src/database/database.py
@@ -65,12 +65,12 @@ from firebase_admin import credentials, firestore
 #     def get_all(self, collection):
 #         """
 #         Retorna todos os documentos de uma coleção.
-        
+
 #         Parameters
 #         ----------
 #         collection : str
 #             Nome da coleção do banco de dados.
-            
+
 #         Returns
 #         -------
 #         list
@@ -83,12 +83,12 @@ from firebase_admin import credentials, firestore
 #     def get_all_keys(self, collection):
 #         """
 #         Retorna todos os IDs dos documentos de uma coleção.
-        
+
 #         Parameters
 #         ----------
 #         collection : str
 #             Nome da coleção do banco de dados.
-        
+
 #         Returns
 #         -------
 #         list
@@ -99,14 +99,14 @@ from firebase_admin import credentials, firestore
 #     def get_by_id(self, collection, id):
 #         """
 #         Retorna um documento do banco de dados.
-        
+
 #         Parameters
 #         ----------
 #         collection : str
 #             Nome da coleção do banco de dados.
 #         id : str
 #             ID do documento a ser retornado.
-        
+
 #         Returns
 #         -------
 #         dict
@@ -124,7 +124,7 @@ from firebase_admin import credentials, firestore
 #             Nome da coleção do banco de dados.
 #         data : dict
 #             Dados a serem adicionados.
-        
+
 #         Returns
 #         -------
 #         str
@@ -135,7 +135,7 @@ from firebase_admin import credentials, firestore
 #     def update(self, collection, id, data):
 #         """
 #         Atualiza um documento do banco de dados.
-        
+
 #         Parameters
 #         ----------
 #         collection : str
@@ -144,11 +144,11 @@ from firebase_admin import credentials, firestore
 #             ID do documento a ser atualizado.
 #         data : dict
 #             Dados a serem atualizados.
-            
+
 #         Returns
 #         -------
 #         bool
-#             True se o documento foi atualizado com sucesso, 
+#             True se o documento foi atualizado com sucesso,
 #             False caso contrário.
 #         """
 #         return self.db.collection(collection).document(id).update(data)
@@ -163,11 +163,153 @@ from firebase_admin import credentials, firestore
 #             Nome da coleção do banco de dados.
 #         id : str
 #             ID do documento a ser deletado.
-        
+
 #         Returns
 #         -------
 #         bool
-#             True se o documento foi deletado com sucesso, 
+#             True se o documento foi deletado com sucesso,
+#             False caso contrário.
+#         """
+#         return self.db.collection(collection).document(id).delete()
+
+
+# class DatabaseManager():
+#     """Singleton para gerenciar o banco de dados"""
+#     _instance = None
+
+#     def __init__(self):
+#         if DatabaseManager._instance is not None:
+#             raise Exception("This class is a singleton!")
+#         else:
+#             cred = credentials.Certificate("serviceAccountKey.json")
+#             firebase_admin.initialize_app(cred)
+#             self.db = firestore.client()
+#             DatabaseManager._instance = self
+
+#     @staticmethod
+#     def instance():
+#         """Retorna a instância do singleton"""
+#         if not DatabaseManager._instance:
+#             DatabaseManager._instance = DatabaseManager()
+#         return DatabaseManager._instance
+
+#     def get_all(self, collection):
+#         """
+#         Retorna todos os documentos de uma coleção.
+
+#         Parameters
+#         ----------
+#         collection : str
+#             Nome da coleção do banco de dados.
+
+#         Returns
+#         -------
+#         dict
+#             Dicionário com os documentos.
+#         """
+#         docs = self.db.collection(collection).get()
+#         return {doc.id: doc.to_dict() for doc in docs if doc.id != "index"}
+
+#     def get_all_keys(self, collection):
+#         """
+#         Retorna todos os IDs dos documentos de uma coleção.
+
+#         Parameters
+#         ----------
+#         collection : str
+#             Nome da coleção do banco de dados.
+
+#         Returns
+#         -------
+#         list
+#             Lista com os IDs dos documentos.
+#         """
+#         docs = self.db.collection(collection).list_documents()
+#         return [doc.id for doc in docs if doc.id != "index"]
+
+#     def get_by_id(self, collection, id):
+#         """
+#         Retorna um documento do banco de dados.
+
+#         Parameters
+#         ----------
+#         collection : str
+#             Nome da coleção do banco de dados.
+#         id : str
+#             ID do documento a ser retornado.
+
+#         Returns
+#         -------
+#         dict
+#             Dicionário com os dados do documento."""
+#         doc = self.db.collection(collection).document(id).get()
+#         return doc.to_dict()
+
+#     def add(self, collection, data):
+#         """
+#         Adiciona um documento ao banco de dados.
+
+#         Parameters
+#         ----------
+#         collection : str
+#             Nome da coleção do banco de dados.
+#         data : dict
+#             Dados a serem adicionados.
+
+#         Returns
+#         -------
+#         str
+#             ID do documento adicionado.
+#         """
+#         # get index counter
+#         index_query = self.db.collection(collection).document("index").get()
+#         if index_query.exists:
+#             index = index_query.to_dict()["index"]
+#             self.db.collection(collection).document("index")\
+#                                           .update({"index": index + 1})
+#         else:
+#             index = 0
+#             self.db.collection(collection).document("index")\
+#                                           .set({"index": 1})
+#         # add document
+#         return self.db.collection(collection).document(str(index)).set(data)
+
+#     def update(self, collection, id, data):
+#         """
+#         Atualiza um documento do banco de dados.
+
+#         Parameters
+#         ----------
+#         collection : str
+#             Nome da coleção do banco de dados.
+#         id : str
+#             ID do documento a ser atualizado.
+#         data : dict
+#             Dados a serem atualizados.
+
+#         Returns
+#         -------
+#         bool
+#             True se o documento foi atualizado com sucesso,
+#             False caso contrário.
+#         """
+#         return self.db.collection(collection).document(id).update(data)
+
+#     def delete(self, collection, id):
+#         """
+#         Deleta um documento do banco de dados.
+
+#         Parameters
+#         ----------
+#         collection : str
+#             Nome da coleção do banco de dados.
+#         id : str
+#             ID do documento a ser deletado.
+
+#         Returns
+#         -------
+#         bool
+#             True se o documento foi deletado com sucesso,
 #             False caso contrário.
 #         """
 #         return self.db.collection(collection).document(id).delete()
@@ -196,12 +338,12 @@ class DatabaseManager():
     def get_all(self, collection):
         """
         Retorna todos os documentos de uma coleção.
-        
+
         Parameters
         ----------
         collection : str
             Nome da coleção do banco de dados.
-            
+
         Returns
         -------
         dict
@@ -213,12 +355,12 @@ class DatabaseManager():
     def get_all_keys(self, collection):
         """
         Retorna todos os IDs dos documentos de uma coleção.
-        
+
         Parameters
         ----------
         collection : str
             Nome da coleção do banco de dados.
-        
+
         Returns
         -------
         list
@@ -230,14 +372,14 @@ class DatabaseManager():
     def get_by_id(self, collection, id):
         """
         Retorna um documento do banco de dados.
-        
+
         Parameters
         ----------
         collection : str
             Nome da coleção do banco de dados.
         id : str
             ID do documento a ser retornado.
-        
+
         Returns
         -------
         dict
@@ -255,7 +397,7 @@ class DatabaseManager():
             Nome da coleção do banco de dados.
         data : dict
             Dados a serem adicionados.
-        
+
         Returns
         -------
         str
@@ -277,7 +419,7 @@ class DatabaseManager():
     def update(self, collection, id, data):
         """
         Atualiza um documento do banco de dados.
-        
+
         Parameters
         ----------
         collection : str
@@ -286,11 +428,11 @@ class DatabaseManager():
             ID do documento a ser atualizado.
         data : dict
             Dados a serem atualizados.
-            
+
         Returns
         -------
         bool
-            True se o documento foi atualizado com sucesso, 
+            True se o documento foi atualizado com sucesso,
             False caso contrário.
         """
         return self.db.collection(collection).document(id).update(data)
@@ -305,11 +447,14 @@ class DatabaseManager():
             Nome da coleção do banco de dados.
         id : str
             ID do documento a ser deletado.
-        
+
         Returns
         -------
         bool
-            True se o documento foi deletado com sucesso, 
+            True se o documento foi deletado com sucesso,
             False caso contrário.
         """
-        return self.db.collection(collection).document(id).delete()
+        if not self.db.collection(collection).document(id).get().exists:
+            return False
+        self.db.collection(collection).document(id).delete()
+        return True
